requirements.txt
colorlog==6.10.1
homeassistant==2026.1.0
pip>=21.3.1
ruff==0.14.10
fiftyfive==0.0.1
python-slugify==8.0.4